+++ src/wallet/test/psbt_wallet_tests.cpp
@@ -5,6 +5,7 @@
 #include <key_io.h>
 #include <script/sign.h>
 #include <util/strencodings.h>
+#include <wallet/psbtwallet.h>
 #include <wallet/rpcwallet.h>
 #include <wallet/wallet.h>
 #include <univalue.h>
